Testes refatoradas.
close #16
